refactor `Role` and ciphertext arithmetic specific to transfer instruction into transfer submodule
